several improvements and fixes
